assignments/assignment_ten/requirements.txt
feedparser
requests
validators
matplotlib
pandas
scipy >= 0.19.0
progressbar2
pandas-ml
seaborn
scikit-learn
numpy
